requirements.txt
# Core libraries
pandas
numpy
matplotlib
pypdf            # or PyPDF2, but not both
python-dotenv
fastapi
uvicorn
langchain
langchain-groq
openai
pdfminer.six
PyMuPDF
streamlit
pydantic>=2.0
python-multipart
boto3
flask
requests
flask_cors      # corrected
openpyxl